fix: replace app/generation.requested with a worker function
